import and export build

+++ src/front/views/Builds/Dashboard/Character/build.tsx
@@ -1,8 +1,10 @@
+import ContentCopyIcon from "@mui/icons-material/ContentCopy";
 import DeleteIcon from "@mui/icons-material/Delete";
-import { Button, Stack, Typography } from "@mui/material";
+import { Button, Stack, Tooltip, Typography } from "@mui/material";
 import type { MouseEvent } from "react";
 import { ElectronEvents } from "src/electron/types";
 import { StackRow } from "src/front/components/Layout/StackRow";
+import { useModalConfirmationContext } from "src/front/components/Modal/Confirmation";
 import { ItemSlot } from "src/front/components/Wakfu/ItemSlot";
 import { useElectronEvent } from "src/front/hooks/electron";
 import { useNavigationContext } from "src/front/views/Navigation";
@@ -18,7 +20,9 @@ export type TCardCharacterBuildProps = {
 
 export const CardCharacterBuild = ({ characterId, build }: TCardCharacterBuildProps) => {
   const { setCurrentView } = useNavigationContext();
+  const confirm = useModalConfirmationContext();
   const [deleteBuild, _, loading] = useElectronEvent(ElectronEvents.BuildDelete);
+  const [serializeBuild] = useElectronEvent(ElectronEvents.BuildSerialize);
 
   const handleClickBuild = () => {
     setCurrentView(NavigationView.BuildDetails, { buildId: build.id });
@@ -26,7 +30,17 @@ export const CardCharacterBuild = ({ characterId, build }: TCardCharacterBuildPr
 
   const handleClickDeleteBuild = (evt: MouseEvent) => {
     evt.stopPropagation();
-    deleteBuild({ characterId, buildId: build.id });
+    confirm("Confirmer la suppression", "Êtes-vous sûr de vouloir supprimer ce build ?").then((result) => {
+      if (result) {
+        deleteBuild({ characterId, buildId: build.id });
+      }
+    });
+  };
+
+  const handleClickExportBuild = async (evt: MouseEvent) => {
+    evt.stopPropagation();
+    const result = await serializeBuild({ buildId: build.id });
+    await navigator.clipboard.writeText(result.serializedBuild);
   };
 
   return (
@@ -39,8 +53,15 @@ export const CardCharacterBuild = ({ characterId, build }: TCardCharacterBuildPr
         {Object.values(EnumWakfuEquipmentPosition).map((position) => (
           <ItemSlot key={position} position={position} item={build.stuff[position]} size={40} disableTooltip />
         ))}
+        <Button variant="push" onClick={handleClickExportBuild}>
+          <Tooltip title="Exporter le build dans le presse-papier" placement="top">
+            <ContentCopyIcon />
+          </Tooltip>
+        </Button>
         <Button variant="push" onClick={handleClickDeleteBuild} disabled={loading} loading={loading}>
-          <DeleteIcon color="error" />
+          <Tooltip title="Supprimer le build" placement="top">
+            <DeleteIcon color="error" />
+          </Tooltip>
         </Button>
       </StackRow>
     </StackRow>
